+++ iphone/Classes/HTTPServerView.mm
@@ -169,7 +169,7 @@ static void *getHostName(void *dummy) {
 }
 
 static void *handle_client_2(void *param) {
-    handle_client((int) param);
+    handle_client((int) (intptr_t) param);
     return NULL;
 }
 
@@ -241,7 +241,7 @@ static void *handle_client_2(void *param) {
         inet_ntop(AF_INET, &ca.sin_addr, cname, sizeof(cname));
         errprintf("Accepted connection from %s\n", cname);
         pthread_t thread;
-        pthread_create(&thread, NULL, handle_client_2, (void *) csock);
+        pthread_create(&thread, NULL, handle_client_2, (void *) (intptr_t) csock);
         pthread_detach(thread);
     }
     
@@ -291,7 +291,7 @@ char *make_temp_file() {
     static int fileno = 0;
     NSString *path = [NSString stringWithFormat:@"%@/tempzip.%d", NSTemporaryDirectory(), ++fileno];
     const char *cpath = [path UTF8String];
-    int len = strlen(cpath);
+    size_t len = strlen(cpath);
     char *ret = (char *) malloc(len + 1);
     strcpy(ret, cpath);
     return ret;
